Rearrange phaser configuration to render within App
