+++ src/Components/App.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+import Game from "./Game";
+
+const App = () => {
+    return(
+        <div>
+            <h2>Hello World</h2>
+            <div id="game-container">
+                <Game/>
+            </div>
+        </div>
+    )
+}
+
+export default App;
